finish

+++ backend/app/controllers/api/v1/users_controller.rb
@@ -3,7 +3,7 @@ class Api::V1::UsersController < ApplicationController
 
   # GET /api/v1/users
   def index 
-    if current_user.admin? 
+    if (current_user.admin?) && (current_user.is_approved?)
       @users = User.all.order(created_at: :desc)
       render json: @users 
     else 
@@ -13,7 +13,7 @@ class Api::V1::UsersController < ApplicationController
 
   # PATCH /api/v1/users/:id/approve
   def approve 
-    if current_user.admin? 
+    if (current_user.admin?) && (current_user.is_approved?)
       user = User.find(params[:id])
       if user.update(is_approved: true)
         render json: { message: "User #{user.username} approved successfully"}
@@ -27,7 +27,10 @@ class Api::V1::UsersController < ApplicationController
 
   # DELETE /api/v1/users/:id
   def destroy 
-    if current_user.admin? 
+    unless current_user.is_approved?
+      return render json: { error: "Your account is pending approval." }, status: :forbidden
+    end
+    if (current_user.admin?) && (current_user.is_approved?)
       user = User.find(params[:id])
       user.destroy
       render json: {message: "User account deleted"}
@@ -37,6 +40,9 @@ class Api::V1::UsersController < ApplicationController
   end 
 
   def update 
+    unless current_user.is_approved?
+      return render json: { error: "Your account is pending approval." }, status: :forbidden
+    end
     user = User.find(params[:id])
     if current_user.id == user.id || current_user.admin? 
       if user.update(user_update_params)
